use more semantically sensible naming with issue.create
And consider then creating a separate function for creating a new
comment in an issue, eg something like
issue.get(issueId).comments.create(text).

And then separate functionality for retrieving and converting the input
from the context passed to the action into issue.Options and
comment.Options.  Which could then be tested separately on unit level,
so that the higher level functionality like issue and comments are
tested on an integration level, and the 2 or 3 user journeys from the
main entrypoint is tested with synthetic transaction in a workflow.