log the data returned by issues.create
In order to create a synthetic transaction which tests the creation of
comments in issues, it would be useful to have access to the
issue_number of a recently created issue so that it can be automatized.
Since all of octokits methods are generated based on a map of all the
endpoints which GitHub provides, then whatever the endpoint for creating
issues returns should also be availble to us. As it turns out it should
return an object containing the issue including the id/issue_number:
https://developer.github.com/v3/issues/#create-an-issue