change use-case for action to create/update issue
For now the update part of upsert is not implemented. Previously though
the use-case for the action was to either create a new issue or create a
comment in the issue. When thinking the concept through there where a
couple of issue with this approach:
- issues and issues comments are sufficiently different that they should
have each their own actions, otherwise the usage will be confusing.
- the use-case of creating an action in a workflow and then creating
comments in the same one, in the same workflow, is both rare and easier
in both implementation and usage to cover with two separate actions.

And on the other hand, creating an action and updating it in the same
workflow is more natural, eg. when adding labels, changing the title,
etc. as things dynamically evolve in a workflow.