validate repository input
any of the following is considered invalid:
- "some-repo"
- ""
- "some-owner/"
- "/some-repo"
- "some-owner/some-owner/some-repo"

it is still possible to receive invalid repository input which still
have both owner and repo but which have any or consists of only
whitespace, which contain any other special disallowed characthers.
Although that seems relatively unlikely, so for now it does not seem
necessary to cover those cases.